refactor(individual-plan): use same component for indivudal plan tab
