+++ src/components/Objective/IndividualPlan/index.ts
@@ -1,0 +1,1 @@
+export { IndividualOkrPage } from './individual-plan'
