+++ src/components/Objective/IndividualPlan/queries.gql
