+++ src/components/Page/MyThings/ActiveCycles/active-cycles.tsx
@@ -4,6 +4,7 @@ import { useRecoilValue, useSetRecoilState } from 'recoil'
 
 import { KeyResultInsertDrawer } from 'src/components/KeyResult/InsertDrawer/wrapper'
 import { KeyResult } from 'src/components/KeyResult/types'
+import { IndividualOkrPage } from 'src/components/Objective/IndividualPlan'
 import { DetailedHeader } from 'src/components/User/DetailedHeader'
 import { keyResultTypeAtom } from 'src/state/recoil/key-result'
 import { keyResultReadDrawerOpenedKeyResultID } from 'src/state/recoil/key-result/drawers/read/opened-key-result-id'
@@ -12,7 +13,6 @@ import meAtom from 'src/state/recoil/user/me'
 import selectUser from 'src/state/recoil/user/selector'
 
 import CompanyOkrPage from './company-okrs'
-import IndividualOkrPage from './individual-okrs'
 
 const ActiveCyclesPage = () => {
   const keyResultType = useRecoilValue(keyResultTypeAtom)
