--- src/components/Page/Profile/queries.gql
@@ -1,26 +1,0 @@
-query LIST_NOT_ACTIVE_CYCLES {
-  cycles(active: false) {
-    edges {
-      node {
-        id
-      }
-    }
-  }
-}
-query LIST_USERS_WITH_INDIVIDUAL_OKR {
-  users(onlyWithIndividualObjectives: true) {
-  edges {
-    node {
-      id
-      firstName
-      nickname
-      fullName
-      role
-      picture
-      about
-      linkedInProfileAddress
-    }
-  }
-  }
-}
-
